Large commit which splits out display functionality into multiple files.
