+++ stimela/backends/kube/kube_utils.py
@@ -5,6 +5,7 @@ from datetime import datetime
 from rich.markup import escape
 from requests import ConnectionError
 from urllib3.exceptions import HTTPError
+from dataclasses import dataclass
 
 from omegaconf import OmegaConf
 
@@ -79,6 +80,24 @@ def apply_pod_spec(kps, pod_spec: Dict[str, Any], predefined_pod_specs: Dict[str
                 log.info(f"setting {kind} CPU limit to {kps.cpu.limit}")
 
     return pod_spec
+
+@dataclass
+class KubeStatusDatum:
+    status = None
+    running_pods = None
+    pending_pods = None
+    terminating_pods = None
+    successful_pods = None
+    failed_pods = None
+    stateless_pods = None
+    total_pods = None
+    total_cores = None
+    total_memory = None
+    connection_status = "connected"
+
+    @property
+    def profiling_results(self):
+        return {"k8s_cores": self.total_cores, "k8s_mem": self.total_memory}
 
 class StatusReporter(object):
     def __init__(self, log: logging.Logger,
@@ -221,7 +240,8 @@ class StatusReporter(object):
         if not self.connected:
             interval = str(datetime.now() - self._last_connected)
             interval = interval.split(".", 1)[0]
-            return [f"lost connection [red]{interval}[/red] ago"], None
+            datum.connection_status = f"disconnected ({interval}s)"
+            return datum
 
         # process metrics if we got them
         if metrics:
@@ -234,50 +254,38 @@ class StatusReporter(object):
                         totals['cpu'] += resolve_unit(usage.get('cpu'), k8s_cpu_units)
                         totals['memory'] += resolve_unit(usage.get('memory'), k8s_memory_units_in_bytes)
                     # print(f"Pod: {pname}, CPU: {usage.get('cpu')}, Memory: {usage.get('memory')}")
+
+        datum = KubeStatusDatum()
+
         # add main pod/job status
-        report_metrics = []
         if self.main_status:
-            report_metrics.append(f"[blue]{self.main_status}[/blue]")
+            datum.status = self.main_status  # Blue
         elif self.podname in self.pod_statuses:
-            report_metrics.append(f"[blue]{self.pod_statuses[self.podname]}[/blue]")
+            datum.status = self.pod_statuses[self.podname]  # Blue
+
         # add count of running pods
-        npods = len(self.pod_statuses)
-        pods = ""
-        nrun = sum([stat.startswith("Running") for stat in self.pod_statuses.values()])
-        if nrun:
-            pods += f"[green]{nrun}[/green]R"
-        npend = sum([stat.startswith("Pending") for stat in self.pod_statuses.values()])
-        if npend:
-            pods += f"[yellow]{npend}[/yellow]P"
-        nterm = sum([stat.startswith("Terminating") for stat in self.pod_statuses.values()])
-        if nterm:
-            pods += f"[blue]{nterm}[/blue]T"
-        nsucc = sum([stat.startswith("Succeeded") for stat in self.pod_statuses.values()])
-        if nsucc:
-            pods += f"[green]{nsucc}[/green]S"
+        datum.total_pods = npods = len(self.pod_statuses)
+        nrun = sum([s.startswith("Running") for s in self.pod_statuses.values()])
+        datum.running_pods = nrun  # Green
+        npend = sum([s.startswith("Pending") for s in self.pod_statuses.values()])
+        datum.pending_pods = npend  # Yellow
+        nterm = sum([s.startswith("Terminating") for s in self.pod_statuses.values()])
+        datum.terminating_pods = nterm  # Blue
+        nsucc = sum([s.startswith("Succeeded") for s in self.pod_statuses.values()])
+        datum.successful_pods = nsucc  # Green
         nfail = sum([stat.startswith("Failed") for stat in self.pod_statuses.values()])
-        if nfail:
-            pods += f"[red]{nfail}[/red]F"
+        datum.failed_pods = nfail # Red
         nuk = npods - nrun - npend - nterm - nsucc - nfail
-        if nuk:
-            pods += f"[red]{nuk}[/red]U"
-        if npods:
-            report_metrics.append(f"pods {pods}")
+        datum.stateless_pods = nuk  # Red
+
         # add metrics
         if metrics:
-            cores = totals['cpu']
-            mem_gb = round(totals['memory'] / 2**30)
-            report_metrics += [
-                f"cores [green]{totals['cpu']:.2f}[/green]",
-                f"mem [green]{mem_gb}[/green]G"
-            ]
-            stats = dict(k8s_cores=cores, k8s_mem=mem_gb)
-        else:
-            stats = None
+            datum.total_cores = totals['cpu']
+            datum.total_memory = round(totals['memory'] / 2**30)
 
         if self._last_disconnected is not None:
-            report_metrics.append("reconnected")
+            datum.connection_status = "connected"
 
-        return report_metrics, stats
+        return datum
