+++ stimela/display/styles/__init__.py
@@ -1,0 +1,4 @@
+from .fancy import FancyDisplay
+from .simple import SimpleDisplay
+from .kube import KubeDisplay
+from .slurm import SlurmDisplay
